Refactor: [FE] utils 폴더 이동 & convert array 함수 추가 및 적용
