+++ client/src/pages/Main/subpages/Expenses/Expenses.infiniteList.tsx
@@ -9,6 +9,8 @@ import { ExpenseType, FormListLayoutType } from "../../../../global/customType";
 import { useRecoilValue } from "recoil";
 import { currentUserAtom } from "../../../../atoms/globalAtoms";
 import Skeleton_ExpenseListItem from "../../../../components/skeleton/Skeleton.expenseListItem";
+import { useGroups } from "../Group/Group.hooks";
+import { covertToStringArray } from "../../../../util/handleCovertArray";
 
 export default function Expenses_List({
   $owner,
@@ -30,6 +32,7 @@ export default function Expenses_List({
   const expenseList = pages.flatMap((page) => page.expenses);
 
   const { showDialog } = useHandleDialog();
+  const { members, categories } = useGroups(currentUser.currentGroup);
 
   function handleClick(
     event: React.SyntheticEvent,
@@ -41,7 +44,11 @@ export default function Expenses_List({
     showDialog({
       type: TYPES.MODAL_DOUBLE_COL, //삭제는 POPUP
       title: LABELS.LABEL_EDIT_EXPENSE,
-      layout: EditExpenseLayout({ item }) as FormListLayoutType[],
+      layout: EditExpenseLayout({
+        item,
+        categories: covertToStringArray(categories, "name"),
+        members: covertToStringArray(members, "nickname"),
+      }) as FormListLayoutType[],
     });
   }
   return (
