+++ client/src/util/handleElement.ts
