+++ client/src/util/updateLocalStorage.ts
